+++ src/main/java/org/example/alphasolutions/controllers/EmployeeController.java
@@ -1,0 +1,17 @@
+package org.example.alphasolutions.controllers;
+
+import org.example.alphasolutions.services.EmployeeService;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/alphaSolutions/emp")
+public class EmployeeController {
+
+    private final EmployeeService employeeService;
+
+    public EmployeeController(EmployeeService employeeService) {
+        this.employeeService = employeeService;
+    }
+
+}
