+++ src/test/java/org/example/alphasolutions/repositories/ManagementSoftwareRepositoryTest.java
@@ -1,4 +1,4 @@
-package org.example.alphasolutions.repositories;
+/*package org.example.alphasolutions.repositories;
 
 import org.example.alphasolutions.models.*;
 import org.junit.jupiter.api.BeforeEach;
@@ -197,4 +197,4 @@ class ManagementSoftwareRepositoryTest {
     }
 
 
-}
+}*/
